EVEREST-483-Edit-backup-schedules: refactoring
