+++ apps/everest/pages/db-cluster-details/backups/scheduled-backup-modal/scheduled-backup-modal-form/scheduled-backup-modal-form.tsx
@@ -18,12 +18,12 @@ import { AutoCompleteInput } from '@percona/ui-lib.form.inputs.auto-complete';
 import { TextInput } from '@percona/ui-lib.form.inputs.text';
 import { LabeledContent } from '@percona/ui-lib.labeled-content';
 import { useParams } from 'react-router-dom';
-import { TimeSelection } from '../../../../components/time-selection/time-selection';
-import { useBackupStorages } from '../../../../hooks/api/backup-storages/useBackupStorages';
-import { useDbCluster } from '../../../../hooks/api/db-cluster/useDbCluster';
-import { Messages } from './scheduled-backup-modal.messages';
-import { ScheduleFields } from './scheduled-backup-modal.types';
-import { ScheduleModalContext } from './context/schedule-modal.context';
+import { TimeSelection } from '../../../../../components/time-selection/time-selection';
+import { useBackupStorages } from '../../../../../hooks/api/backup-storages/useBackupStorages';
+import { useDbCluster } from '../../../../../hooks/api/db-cluster/useDbCluster';
+import { Messages } from '../scheduled-backup-modal.messages';
+import { ScheduleFields } from '../scheduled-backup-modal.types';
+import { ScheduleModalContext } from '../context/schedule-modal.context';
 import { useFormContext } from 'react-hook-form';
 
 export const ScheduledBackupModalForm = () => {
